# refactor-baseline/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "wardenIQ \u2014 Test Intelligence Platform",
    "version": "0.1.0-beta"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/features/{fid}/jira-sync": {
      "post": {
        "description": "Write the feature's current coverage + open PRs back to its Jira issue.",
        "operationId": "jira_sync_api_features__fid__jira_sync_post",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Jira Sync"
      }
    },
...
    "/api/projects/{pid}/jira-issues": {
      "get": {
        "description": "Epics available to associate with a feature: every Epic in the linked Jira\nproject MINUS any Epic already bound to a feature (1 Epic : 1 feature).\n`exclude_feature_id` keeps the epic already bound to that feature visible\n(so its own selection shows when editing / re-versioning).",
        "operationId": "list_project_jira_issues_api_projects__pid__jira_issues_get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "exclude_feature_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Exclude Feature Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Jira Issues"
      }
    },
...
  }
}